Update MergeData.R
Trying to get box plot working. Found the problem but not the solution. Problem is the Rotten Tomato score column is chr instead of a int/float. Have to convert chr to int/float. I tried as.numeric however since input is "30/100"  fractions, r does not convert to a number.